Support rendering pages without the nav and footer for iframes.
LoudSwarm wants to have links to various types of content that
exists on our site already. Rather than duplicating and dealing
with stale data, this allows our site to be more easily used in
an iframe.

This relies on JavaScript to detect that we're in an iframe, then
remove the bulky nav and footer elements. This means that if a user
has JavaScript disabled (they're going to have a tough time on LoudSwarm)
they would still see the header and footer. Additionally, if they are
using a slower machine, the header and footer would remain visible for
a noticeable amount of time. But again, they're probably going to have
a bad time on the LoudSwarm platform.