Allow wrapping of the images on the volunteer post.
The nowrap spans are so that we don't end up with a single word next
to the image on mobile. It was occuring with we, but I figured I should
do it for all of them.

This also fixes an issue with the wrong image name being specified and
utilizes volunteer4.jpg for the part that talks about working the
registration desk.